chore: remove obsolete hooks and references
